apps/api/pyproject.toml
[project]
name = "sanity"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
dev = "uvicorn sanity.main:app --reload --host 127.0.0.1 --port 8000"
format = "ruff format ."
lint = "ruff check ."
lint_fix = "ruff check --fix ."
db_migrate = "uv run alembic upgrade head"
db_downgrade = "uv run alembic downgrade -1"
seed_catalogue = "uv run python -m scripts.seed_catalogue"

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"migrations/versions/*.py" = [
    "E501" # Line too long
]
